overhauled UI
the UI was looking like it was from the 90s so I overhauled it, taking about 4 hours because I had a lot of bugs along the way. The pom.xml file had a dependency that I thought it needed that it apparently shouldn't have had. Also, once the UI was better, it was also taking more resources, which made the timer giltch in its updating every second, which I had to learn how to fix. At least it's prettier now