Broke the dropdown and main play button, but at least I added some comments
I'm just updating because it has been a while. I'm not sure when but at some point I broke both the dropdown and the play button (the play button was working last night so I don't know what happened)
I've been trying to find the issue but no luck yet. I added a few comments to make my code more readable and I plan to continue with that.